fix(): loader should not fetch sourcemaps if disabled.
* Currently if the trace option `sourceMaps` is set to false, the sourceMaps will be still loaded.
  This isn't valid. Once the option is set to false, it should just ignore the sourceMaps.

Fixes #610.
