+++ lib/builder.js
@@ -5,8 +5,6 @@ var asp = require('bluebird').promisify;
 var fs = require('fs');
 var path = require('path');
 
-var dataUriToBuffer = require('data-uri-to-buffer');
-
 var extend = require('./utils').extend;
 
 var attachCompilers = require('./compile').attachCompilers;
@@ -190,29 +188,7 @@ Builder.prototype.reset = function(baseLoader) {
           throw err;
         });
       })
-    }))
-
-    // support picking up source URLs during fetch to set as metadata.sourceMap
-    .then(function (source) {
-      if (load.metadata.sourceMap)
-        return source;
-      var sourceMap = source.match(/^\s*\/\/\s*[#@] sourceMappingURL=([^\s'"]*)/m);
-      if (!sourceMap)
-        return source;
-      if (sourceMap[1].startsWith('data:')) {
-        // inline source map
-        load.metadata.sourceMap = JSON.parse(dataUriToBuffer(sourceMap[1]).toString('utf8'));
-        return source;
-      }
-      else {
-        // relative path to the .map file
-        var sourceMapPath = path.join(path.dirname(fromFileURL(load.address)), sourceMap[1]);
-        return asp(fs.readFile)(sourceMapPath, 'utf8').then(function (sourceMap) {
-          load.metadata.sourceMap = JSON.parse(sourceMap);
-          return source;
-        });
-      }
-    });
+    }));
   };
 
   // this allows us to normalize package conditionals into package conditionals
